user-service/requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de données
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Validation et sérialisation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP et requêtes
httpx==0.25.2
requests==2.31.0

# Utilitaires
python-dotenv==1.0.0
email-validator==2.1.0

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Développement
black==23.11.0
isort==5.12.0
flake8==6.1.0
